ICD10Utilities.jl

This package provide functions for working with ICD-10 codes.

It is intended for researchers working with ICD-10 codes in administrative health data. This is not a package for clinical coding.

Note

There are no ICD codes included in this package.

ICD-10

The International Classification of Diseases and Related Health Problems is more than 100 years old. The tenth revision, International Classification of and Related Health Problems, 10th Revision (ICD-10) was first released by the World Health Organization (WHO) in 1992. It is used for classifying diagnoses in health data.

Some countries have developed customised versions under licence from WHO. These include Australia (ICD-10-AM), Canada (ICD-10-CA), the United States (ICD-10-CM), and Germany (ICD-10-GM). This package defines types for each of these customised versions, and functions for working with them.

Use

An ICD-10 code can be created from a string using ICD10("A00.1"). Similar constructors exist for the country-specific versions (e.g. ICD10AM("A00.1")).

Some databases store ICD-10 codes as strings without the .. In this case, you can pass a second argument false to indicate that punctuation is not used (e.g. ICD10AM("A001", false)). If you do not set the second argument correctly for your input, you wiil get wrong codes as output.

You can also validate the input format by passing a third argument (e.g. ICD10CA("A00.1", true, true)). This will usually not be necessary as in most cases codes will be obtained from a health database and so will already have been validated. It is much slower to validate the input format.

Types

ICD10Utilities.ICD10 — Type

ICD10

An ICD-10 code.

ICD10Utilities.ICD10AM — Type

ICD10AM

Australian version of ICD-10

ICD10Utilities.ICD10CA — Type

ICD10CA

Canadian version of ICD-10

ICD10Utilities.ICD10CM — Type

ICD10CM

United States version of ICD-10

ICD10Utilities.ICD10GM — Type

ICD10GM

German version of ICD-10

Functions

ICD10Utilities.ICD10 — Method

ICD10(ANNx::String, punct=true, validateinput = false)

Create an ICD-10 code from a string.

ICD-10 codes have the form ANN[.][N[N]] where A is a letter A-Z and N is a digit 0-9. Parts in brackets are optional.

If punct=false then the . is assumed not present.

If validateinput=true then the input string is checked for valid format using a regex. This takes 2–3 times as long as not checking.

ICD10Utilities.ICD10 — Method

ICD10(icd::T) where T <: AbstractICD10

Create an ICD-10 code from another type of ICD-10 code.

Note this just changes the type. It does not translate concepts between versions.

ICD10Utilities.ICD10AM — Method

ICD10AM(ANNx::String, punct=true, validateinput=false)

Create an Australian ICD-10-AM code from a string.

If punct=false then the string is assumed not to contain a .

If validateinput=true then the input format is checked using a regex. this takes 2–3 times as long as not validating.

ICD10Utilities.ICD10AM — Method

ICD10AM(icd::T) where {T<:AbstractICD10}

Convert ICD-10 code to Australian ICD-10-AM.

Note this just changes the type of the code. It does not do any checking or translating of concepts between versions.

ICD10Utilities.ICD10CA — Method

ICD10CA(ANNx::String, punct=true, validateinput=false)

Create a Canadian ICD-10-CA code from a string.

If punct=false then the string is assumed not to contain a .

If validateinput=true then the input format is checked using a regex. this takes 2–3 times as long as not validating.

ICD10Utilities.ICD10CA — Method

ICD10CA(icd::T) where {T<:AbstractICD10}

Convert ICD-10 code to Canadian ICD-10-CA.

Note this just changes the type of the code. It does not do any checking or translating of concepts between versions.

ICD10Utilities.ICD10CM — Method

ICD10CM(ANNx::String, punct=true, validateinput=false)

Create a United States ICD-10-CM code from a string.

If punct=false then the string is assumed not to contain a .

If validateinput=true then the input format is checked using a regex. this takes 2–3 times as long as not validating.

ICD10Utilities.ICD10CM — Method

ICD10CM(icd::T) where {T<:AbstractICD10}

Convert ICD-10 code to United States ICD-10-CM.

Note this just changes the type of the code. It does not do any checking or translating of concepts between versions.

ICD10Utilities.ICD10GM — Method

ICD10GM(ANNx::String, punct=true, validateinput=false)

Create a German ICD-10-GM code from a string.

If punct=false then the string is assumed not to contain a .

If validateinput=true then the input format is checked using a regex. this takes 2–3 times as long as not validating.

ICD10Utilities.ICD10GM — Method

ICD10GM(icd::T) where {T<:AbstractICD10}

Convert ICD-10 code to German ICD-10-GM.

Note this just changes the type of the code. It does not do any checking or translating of concepts between versions.

Base.string — Method

string(icdcode)
string(icdcode, false)

Convert an ICD10 code to a string. Output will have punctuation unless false is passed as second argument.

ICD10Utilities.isvalidcode — Method

isvalidcode(icd, validcodes)

Takes an AbstractICD10 and compares it with a list of valid codes.