ICD-10-AM/ACHI

ICD-10-AM is the Australian modification of ICD-10. It has been in use for coding diagnoses in all Australian public and private hospitals since 1998.

The Australian Classification of Health Interventions is a companion coding system for coding medical procedures and other interventions in Australia.

ICD-10-AM/ACHI is updated every 2–3 years. THe current 11th edition has been in use since 1 July 2019.

ICD-10-AM/ACHI and the Australian Coding Standards (ACS) are available for purchase from the Independent Hospital Pricing Authority (IHPA).

If you have access to the ICD-10-AM/ACHI electronic code lists (from IHPA), you can import them into this package (see ICD10Utilities.importicd10amcodes and ICD10Utilities.importachicodes ) and use them for validating codes and conducting other queries (e.g. look up description from a code). See ICD-10-AM code list and ACHI code list.

Functions

ICD-10-AM

ICD10Utilities.importicd10amcodes — Function

importicd10amcodes(fname)

Import the ICD-10-AM electronic code lists into the package.

Electronic code lists can be obtained from the Independent Hospital Pricing Authority.

This function just imports the disease.txt table. Provide the path to the disease.txt file.

ICD10Utilities.isvalidcode — Method

isvalidcode(icdcode::ICD10AM)

Takes an ICD10AM code and checks it is in the electronic code lists and valid.

Electronic code lists can be obtained from the Independent Hospital Pricing Authority, and need to be imported into the local package before this function can be used. See importicd10amcodes

"Valid" codes are those that exist and are valid for use in coding (e.g. 3-digit codes are often not valid for use if they have 4- or 5-digit subcodes).

ACHI

ICD10Utilities.ACHI — Type

ACHI

An ACHI (Australian Classification of Health Interventions) code.

ICD10Utilities.ACHI — Method

ACHI(str::String, validateinput=false)

Create an ACHI code from a string.

ACHI codes consist of 7 digits (0-9), usually with a hypen after the first 5 digits.

If validateinput=true then the input is checked using a regex. Note this only checks the input format; it does not check if the code actually exists in ACHI.

ICD10Utilities.importachicodes — Function

importachicodes(fname)

Import the ACHI electronic code lists into the package.

Electronic code lists can be obtained from the Independent Hospital Pricing Authority.

This function just imports the interven.txt table. Provide the path to the interven.txt file.

ICD10Utilities.isvalidcode — Method

isvalidcode(code:ACHI)

Takes an ACHI code and checks it is in the electronic code lists and valid.

Electronic code lists can be obtained from the Independent Hospital Pricing Authority, and need to be imported into the local package before this function can be used. See importachicodes

"Valid" codes are those that exist.

Code lists

Note

These tables are only available if you have imported the electronic code lists obtained from IHPA.

ICD-10-AM code list

ICD10Utilities.ICD10AMcodes

The imported ICD-10-AM electronic code list.

This is a TypedTable created from the disease.txt table in the electronic code lists. Some variable names are changed.

ICD10AMcodes	disease.txt	description

level	Level	Number of digits in the code (3, 4, 5)

icdcode	code_id	ICD-10-AM code

dagger	dagger	Is this a dagger code

asterisk	asterisk	Is this an asterisk code

validforcoding	valid	Can this code be used in coding, or is it just a placeholder for subcodes

australiancode	australian	Is this an Australian code (not part of WHO ICD-10)

description	ascii_desc	Description of the coded concept

shortdescription	ascii_short_desc	Abbreviated description

effectivefrom	effective_from	Date the code came into use

inactive	inactive	Date when code became inactive

reactivated	reactivated	Date when code was reactivated

sex	Sex	Is code sex specific

stype	Stype

agel	AgeL	Minimum patient age for assigning the code

ageh	AgeH	Maximum patient age for assigning the code

atype	Atype

rarenotifiabledx	RDiag	Code refers to a rare or notifiable disease

morphcoderequired	Morph_Code	Use of this code requires an accompanying morphology code

conceptchange	concept_change	Date this code was introduced to represent a new concept

unacceptpdx	UnacceptPDx	Can this code be used as a principal diagnosis

ACHI code list

ICD10Utilities.ACHIcodes

The imported ACHI electronic code list.

This is a TypedTable created from the interven.txt table in the electronic code lists. Some variable names are changed.

ACHIcodes	interven.txt	description

achicode	Code_id	ACHI code

block	Block	ACHI block that includes the code

description	ascii_desc	Description of the intervention

shortdescription	ascii_short_desc	Abbreviated description

effectivefrom	effective_from	Date the code came into use

inactive	inactive	Date the code became inactive

sex	Sex	Code is sex-specific

stype	Stype

agel	AgeL	Minimum patient age for assigning this code

ageh	AgeH	Maximum patient age for assigning this code

atype	Atype