ICD-10-CM

ICD-10-CM is the United States Clinical Modification of ICD-10.

It differs from other version in allowing up to 7 digits in codes, and allowing letters as well as numbers in digits 3–7.

Types

ICD10Utilities.ICD10CM — Type

ICD10CM

An ICD-10-CM code

Fields are level, indicating the number of digits in the code, and data containing the code.

Functions

ICD10Utilities.ICD10CM — Method

ICD10CM(str::String, validateinput = false)

Create an ICD10CM code from a string.

ICD10CM codes have the form ANC[.][C[C[C[C]]]] where A is a letter A-Z, N is a digit 0-9, and C is a letter or digit. Parts in brackets are optional.

If validateinput=true then the input string is checked for valid format using a regex.

ICD10Utilities.ICD10CM — Method

ICD10CM(icd::T) where T <: AbstractICD10

Create an ICD-10-CM code from another type of ICD-10 code.

Note this just changes the type. It does not translate concepts between versions.

Most ICD-10 versions have up to 5 digits, while ICD-10-CM has up to 7, so this conversion cannot be exact.

ICD10Utilities.ICD10 — Method

ICD10(icd::ICD10CM)

Convert from ICD10CM to ICD10.

Note that ICD-10-CM codes may have up to 7 digits, while other ICD-10 versions only have up to 5.

This function truncates longer codes.

Note this just changes the type. It does not translate concepts between versions.

ICD10Utilities.ICD10AM — Method

ICD10AM(icd::ICD10CM)

Convert from ICD10CM to ICD10AM.

Note that ICD-10-CM codes may have up to 7 digits, while other ICD-10 versions only have up to 5.

This function truncates longer codes.

Note this just changes the type. It does not translate concepts between versions.

ICD10Utilities.ICD10CA — Method

ICD10CA(icd::ICD10CM)

Convert from ICD10CM to ICD10CA.

Note that ICD-10-CM codes may have up to 7 digits, while other ICD-10 versions only have up to 5.

This function truncates longer codes.

Note this just changes the type. It does not translate concepts between versions.

ICD10Utilities.ICD10GM — Method

ICD10GM(icd::ICD10CM)

Convert from ICD10CM to ICD10GM.

Note that ICD-10-CM codes may have up to 7 digits, while other ICD-10 versions only have up to 5.

This function truncates longer codes.

Note this just changes the type. It does not translate concepts between versions.