ICD-O-3

The International Classification of Diseases for Oncology, 3rd edition defines codes for describing cancer morphology.

These codes consist of the letter 'M', followed by 4 decimal digits describing the histology, a slash '/', and a single decimal digit describing the behaviour.

Types

ICD10Utilities.ICDO3 — Type

ICDO3

An ICD-O-3 cancer morphology code.

Functions

ICD10Utilities.ICDO3 — Method

ICDO3(str::String, validateinput=false)

Create an ICDO3 code from a string.

ICDO3 codes consist of the letter M followed by 4 digits, a slash /, and a final behaviour digit.

If validateinput=true then the input is checked using a regex. Note this only checks the input format; it does not check if the code actually exists.