ICD10Utilities.jl

This package provide functions for working with ICD-10 codes.

It is intended for researchers working with ICD-10 codes in administrative health data. This is not a package for clinical coding.

Note

There are no ICD codes included in this package.

ICD-10

The International Classification of Diseases and Related Health Problems is more than 100 years old. The tenth revision, International Classification of and Related Health Problems, 10th Revision (ICD-10) was first released by the World Health Organization (WHO) in 1992. It is used for classifying diagnoses in health data.

Some countries have developed customised versions under licence from WHO. These include Australia (ICD-10-AM), Canada (ICD-10-CA), the United States (ICD-10-CM), and Germany (ICD-10-GM). This package defines types for each of these customised versions, and functions for working with them.

Use

An ICD-10 code can be created from a string using ICD10("A00.1"). Similar constructors exist for the country-specific versions (e.g. ICD10AM("A00.1")).

Some databases store ICD-10 codes as strings without the .. The constructor will recognise this and correctly parse the string as an ICD-10 code.

You can also validate the input format by passing a second argument (e.g. ICD10CA("A00.1", true)). This will usually not be necessary as in most cases codes will be obtained from a health database and so will already have been validated. It is slower to validate the input format.

Types

ICD10Utilities.AbstractICD10 — Type

AbstractICD10

Abstract type for ICD-10 codes

ICD10Utilities.ICD10 — Type

ICD10

An ICD-10 code (standard WHO version).

Fields are level, indicating if it is a 3, 4, or 5 digit code, and data containing the code.

ICD10Utilities.ICD10AM — Type

ICD10AM

Australian version of ICD-10

ICD10Utilities.ICD10CA — Type

ICD10CA

Canadian version of ICD-10

ICD10Utilities.ICD10GM — Type

ICD10GM

German version of ICD-10

Functions

ICD10Utilities.ICD10 — Method

ICD10(str::String, validateinput = false)

Create an ICD10 code from a string.

ICD10 codes have the form ANN[.][N[N]] where A is a letter A-Z and N is a digit 0-9. Parts in brackets are optional.

If validateinput=true then the input string is checked for valid format using a regex.

ICD10Utilities.ICD10 — Method

ICD10(icd::T) where T <: AbstractICD10

Create an ICD-10 code from another type of ICD-10 code.

Note this just changes the type. It does not translate concepts between versions.

ICD10Utilities.ICD10AM — Method

ICD10AM(str::String, validateinput = false)

Create an ICD10AM code from a string.

ICD10AM codes have the form ANN[.][N[N]] where A is a letter A-Z and N is a digit 0-9. Parts in brackets are optional.

If validateinput=true then the input string is checked for valid format using a regex.

ICD10Utilities.ICD10AM — Method

ICD10AM(icd::T) where {T<:AbstractICD10}

Convert ICD-10 code to Australian ICD-10-AM.

Note this just changes the type of the code. It does not do any checking or translating of concepts between versions.

ICD10Utilities.ICD10CA — Method

ICD10CA(str::String, validateinput = false)

Create an ICD10CA code from a string.

ICD10CA codes have the form ANN[.][N[N]] where A is a letter A-Z and N is a digit 0-9. Parts in brackets are optional.

If validateinput=true then the input string is checked for valid format using a regex.

ICD10Utilities.ICD10CA — Method

ICD10CA(icd::T) where {T<:AbstractICD10}

Convert ICD-10 code to Canadian ICD-10-CA.

Note this just changes the type of the code. It does not do any checking or translating of concepts between versions.

ICD10Utilities.ICD10GM — Method

ICD10GM(str::String, validateinput = false)

Create an ICD10GM code from a string.

ICD10GM codes have the form ANN[.][N[N]] where A is a letter A-Z and N is a digit 0-9. Parts in brackets are optional.

If validateinput=true then the input string is checked for valid format using a regex.

ICD10Utilities.ICD10GM — Method

ICD10GM(icd::T) where {T<:AbstractICD10}

Convert ICD-10 code to German ICD-10-GM.

Note this just changes the type of the code. It does not do any checking or translating of concepts between versions.

ICD10Utilities.isvalidcode — Method

isvalidcode(icd, validcodes)

Takes an AbstractICD10 and compares it with a list of valid codes.

ICD10Utilities.icd3 — Method

icd3(icdcode)

Returns the initial 3 digits of an ICD10 code. Use for comparing codes just on first 3 digits.

ICD10Utilities.icd4 — Method

icd4(icdcode)

Returns the initial 4 digits of an ICD10 code. Use for comparing codes just on first 4 digits.

ICD10Utilities.seticdpunct — Method

seticdpunct(v::Bool)

Set ICDOPTS[:punct] to new value.

Options

ICD10Utilities.ICDOPTS — Constant

ICDOPTS

Dictionary of options for ICD10Utilities.

:punct: if true (default), ICD-10 codes are displayed with . after third character, and ACHI codes are displayed with - after fifth character.